api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
torch==2.2.0
diffusers==0.25.0
Pillow==10.2.0
python-multipart==0.0.7 
requests>=2.31.0
boto3>=1.34.0
python-dotenv>=1.0.0
supabase>=2.3.0
tk>=0.1.0
requests
fastapi
uvicorn
python-dotenv
boto3
supabase 